+++ app/(home)/top/page.tsx
@@ -6,12 +6,12 @@ import SidebarLeft from "../../../components/home/SidebarLeft";
 import MainContent from "../../../components/home/MainContent";
 import SidebarRight from "../../../components/home/SidebarRight";
 import TweetFrom from "../../../components/home/TweetFrom";
+
 import '../../../static/css/app.css';
 
 export default function Top() {
 
   const [isTweetFormModalOpen, setIsTweetFormModalOpen] = useState(false);
-  const [toast, setToast] = useState("");
 
   return(
     <div className="grid grid-container bg-gray-900 text-white">
